Reorganized upgrade classes into include/datafile
