+++ src/core/UpgradeExtendedNoteRange.cpp
@@ -21,7 +21,7 @@
  *
  */
 
-#include "UpgradeExtendedNoteRange.h"
+#include "datafile/UpgradeExtendedNoteRange.h"
 
 #include "Track.h"
