+++ src/core/UpgradeRenameBBTCO.cpp
@@ -22,7 +22,7 @@
  *
  */
 
-#include "UpgradeRenameBBTCO.h"
+#include "datafile/UpgradeRenameBBTCO.h"
 
 #include <vector>
