+++ src/core/UpgradeTo1_3_0.cpp
@@ -21,7 +21,7 @@
  *
  */
 
-#include "UpgradeTo1_3_0.h"
+#include "datafile/UpgradeTo1_3_0.h"
 
 #include <algorithm>
 #include <cmath>
